win32: remove -municode from mpv binary
The new build system emits a main function when performing configure
tests that require linking. If -municode is specified, this always
fails, because the runtime looks for wmain instead of main. This is a
big mess and would require windows-specific hacks in the build system.

Avoid this by always using main(), and calling GetCommandLineW()
instead. This function returns a single string, so use
CommandLineToArgvW() to split it, and hope it has the same semantics.
Should this ever return NULL, hope that it leaves argc at 0.