command: make vf-metadata/af-metadata somewhat observable
Until now they weren't observable and never reported any updates. Apply
a shitty hack to make them mostly-observable. It relies on the "idle"
event, which is basically triggered on every frame displayed, or
similar. This can lead to property change notifications not being sent
quickly enough.

The cleaner solution would be adding a notification mechanisms from
filters, but I'm too lazy for that.